server/requirements.txt
django-environ
asgiref
backports.zoneinfo==0.2.1
Django==4.1.2
django-environ==0.9.0 # This
psycopg2-binary==2.9.4
sqlparse==0.4.3
djangorestframework
dj-rest-auth
django-allauth
gunicorn